+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/AwesomeBarFeature.kt
@@ -57,9 +57,10 @@ class AwesomeBarFeature(
      */
     fun addSearchProvider(
         searchEngine: SearchEngine,
-        searchUseCase: SearchUseCases.DefaultSearchUseCase
+        searchUseCase: SearchUseCases.DefaultSearchUseCase,
+        mode: SearchSuggestionProvider.Mode = SearchSuggestionProvider.Mode.SINGLE_SUGGESTION
     ): AwesomeBarFeature {
-        awesomeBar.addProviders(SearchSuggestionProvider(searchEngine, searchUseCase))
+        awesomeBar.addProviders(SearchSuggestionProvider(searchEngine, searchUseCase, mode))
         return this
     }