+++ mobile/android/android-components/docs/changelog.md
@@ -24,6 +24,8 @@ permalink: /changelog/
       )
     }
   ```
+* **feature-awesomebar**
+  * Added ability to show one item per search suggestion ([#1779](https://github.com/mozilla-mobile/android-components/issues/1779))
 
 * **lib-publicsuffixlist**
   * The public suffix list shipping with this component is now updated automatically in the repository every day (if there are changes).